Generalization of flotationFoam code
A flotation model is now associated with a two-phase interface. The flotation
model is then applied to both phases, describing the exchange of particles
between both phases. Source terms that account for particle exchange between
phases can be selected in the same way as in IATE and LogMoM, i.e., as a list of
sources. Before, there was an asymmetry in the flotation model in the sense that
there was a 'free' particle phase and a 'captured' particle phase. This is no
longer the case. Whether particles stay in one phase or more to another is fully
determined by the selection of source terms.